Refactor phone number validation
