+++ apps/lv-web/src/app/dashboard/profile/page.tsx
@@ -3,38 +3,17 @@
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { Input } from '@/components/ui/input';
+import {
+  isPhoneCharactersValid,
+  normalizePhoneForDisplay,
+  normalizePhoneForStorage,
+  PROFILE_FIELD_LIMITS,
+} from '@/lib/profile-validation';
 import { useSession } from 'next-auth/react';
 import { useRouter } from 'next/navigation';
 import { useEffect, useState } from 'react';
 import { toast } from 'sonner';
 import { ProfileGetResponse, UserProfile } from '../../api/profile/route';
-
-const PROFILE_FIELD_LIMITS = {
-  firstName: 50,
-  lastName: 50,
-  displayName: 100,
-  phone: 20,
-  bio: 500,
-} as const;
-
-const PHONE_ALLOWED_CHARS_REGEX = /^[+\-()0-9]*$/;
-
-const normalizePhoneInput = (value: string) => {
-  return value.replace(/[^+\-()0-9]/g, '').slice(0, PROFILE_FIELD_LIMITS.phone);
-};
-
-const normalizePhoneForDisplay = (value: string | null | undefined) => {
-  if (!value) return '';
-
-  return normalizePhoneInput(value);
-};
-
-const normalizePhoneForStorage = (value: string | null | undefined) => {
-  if (!value) return null;
-
-  const normalizedPhone = normalizePhoneInput(value);
-  return normalizedPhone === '' ? null : normalizedPhone;
-};
 
 export default function ProfilePage() {
   const { data: session, status } = useSession();
@@ -87,7 +66,7 @@ export default function ProfilePage() {
     try {
       const normalizedPhone = normalizePhoneForStorage(profile.phone);
 
-      if (normalizedPhone && !PHONE_ALLOWED_CHARS_REGEX.test(normalizedPhone)) {
+      if (!isPhoneCharactersValid(normalizedPhone)) {
         toast.error('Phone number can only contain +, -, (, ) and digits');
         return;
       }
